Add test suites for ruby of credential valiator new test suites for testing seperate module credentials validator, which is for creating new object for validation and validating this object
