+++ spec/controllers/ems_middleware_controller_spec.rb
@@ -7,7 +7,7 @@ describe EmsMiddlewareController do
     controller.instance_variable_set(:@breadcrumbs, [])
     get :new
 
-    expect(response.status).to eq(200)
+    expect(response.status).to eq(302)
     expect(allow(controller).to receive(:edit)).to_not be_nil
   end
